pydantic-ai-agent/requirements.txt
pydantic-ai-slim[google,ag-ui]
fastapi
uvicorn[standard]
python-dotenv
